+++ lib/rubocop/cop/minitest_cops.rb
@@ -1,4 +1,5 @@
 # frozen_string_literal: true
 
 require_relative 'minitest/assert_nil'
+require_relative 'minitest/assert_truthy'
 require_relative 'minitest/refute_nil'